requirements.txt
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
duckdb==1.0.0
sqlalchemy==2.0.31
requests==2.32.3
kaggle==1.6.14
pyarrow==16.1.0
papermill==2.6.0
jupyter==1.0.0
notebook==7.2.1
plotly==5.23.0
matplotlib==3.9.0
scikit-learn==1.5.1